fix: jsonrpc request parsing error
"Content-Type" in jsonrpc request indicates the byte length of the
following json string. If the json string contains multi-byte
characters, the json string obtained by use "String.slice" is longer
than the actual json string, which will lead to the failure of parsing
the json string into a json object.